# Conflicts:
#	composer.lock
#	src/BeneficiaryBundle/Utils/BeneficiaryService.php
#	src/CommonBundle/Controller/ExportController.php
#	src/CommonBundle/Resources/config/services.yml
#	src/DistributionBundle/Controller/AssistanceController.php
#	src/DistributionBundle/Resources/config/services.yml
#	src/DistributionBundle/Utils/AssistanceService.php
#	src/NewApiBundle/Controller/BookletController.php
#	src/NewApiBundle/Controller/CommonController.php
#	src/NewApiBundle/Controller/CountrySpecificController.php
#	src/NewApiBundle/Controller/DonorController.php
#	src/NewApiBundle/Controller/OrganizationController.php
#	src/NewApiBundle/Controller/ProductController.php
#	src/NewApiBundle/Resources/config/services.yml
#	src/ProjectBundle/Utils/ProjectService.php
#	src/VoucherBundle/Controller/SmartcardController.php
#	src/VoucherBundle/Utils/BookletService.php
#	src/VoucherBundle/Utils/SmartcardService.php
#	tests/VoucherBundle/Controller/SmartcardControllerTest.php

+++ src/DistributionBundle/Controller/AssistanceController.php
@@ -26,6 +26,7 @@ use Nelmio\ApiDocBundle\Annotation\Model;
 use Swagger\Annotations as SWG;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\ParamConverter;
+use Symfony\Component\HttpKernel\Exception\BadRequestHttpException;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Symfony\Component\Serializer\SerializerInterface;
 
@@ -341,7 +342,15 @@ class AssistanceController extends Controller
     {
         $deletionData = $request->request->all();
 
-        $return = $this->assistanceBeneficiaryService->removeBeneficiaryInDistribution($distribution, $beneficiary, $deletionData);
+        if (true === $distribution->getCompleted() || true === $distribution->getArchived()) {
+            throw new BadRequestHttpException("Beneficiary can't be removed from closed or archived assistance");
+        }
+
+        try {
+            $return = $this->assistanceBeneficiaryService->removeBeneficiaryInDistribution($distribution, $beneficiary, $deletionData);
+        } catch (\InvalidArgumentException $e) {
+            throw new BadRequestHttpException($e->getMessage(), $e);
+        }
 
         return new Response(json_encode($return));
     }
